Cleared up problem with blank combobox popups after adding flight configurations.
(1) Created new ConfigurationComboBox, extended from
JComboBox<FlightConfiguration>, with a listener for popup opens that
connects to a new ConfigurationModel.

(2) Removed some cruft from ConfigurationModel and made it a nested
class within ConfigurationComboBox.

(3) Updated ComponentAnalysisDialog, RocketPanel, and
SimulationEditDialog to use ConfigurationComboBox.
